+++ app/Unit/Summary/UnitViewSummary.jsx
@@ -5,38 +5,25 @@ import {
 } from 'react-router-dom';
 import {connect} from "react-redux";
 import classnames from 'classnames';
-import MarksArticle from './MarksArticle.jsx';
-import {NodesExtensible} from './NodesDisplay/NodesDisplay.jsx';
-import AuthorStatics from './Author/AuthorStatics.jsx';
-import DateConverter from '../Component/DateConverter.jsx';
-import {AccountPlate} from '../Component/AccountPlate.jsx';
+import MarksArticle from '../MarksArticle.jsx';
+import {NodesExtensible} from '../NodesDisplay/NodesDisplay.jsx';
+import ActionBroad from '../Actions/ActionBroad.jsx';
+import DateConverter from '../../Component/DateConverter.jsx';
+import {AccountPlate} from '../../Component/AccountPlate.jsx';
 
 const styleMiddle = {
-  boxStatics: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'space-between',
-    width: '8%',
-    height: '64%',
-    position: 'absolute',
-    bottom: '13%',
-    left: '36%',
-    boxSizing: 'border-box'
-  },
+
 }
 
-class UnitAuthorSummary extends React.Component {
+class UnitViewSummary extends React.Component {
   constructor(props){
     super(props);
     this.state = {
-      onActEdit: false
+
     };
     this.marksArticle = React.createRef();
     this._set_layerstatus = this._set_layerstatus.bind(this);
     this._handleClick_Account = this._handleClick_Account.bind(this);
-    this._handleEnter_actEdit = this._handleEnter_actEdit.bind(this);
-    this._handleLeave_actEdit = this._handleLeave_actEdit.bind(this);
-    this._handleClick_UnitAction_Author = this._handleClick_UnitAction_Author.bind(this);
     this._handleWheel_marksArticle = (event)=>{event.stopPropagation();};
     this.style={
       Com_UnitViewSummary_: {
@@ -48,7 +35,7 @@ class UnitAuthorSummary extends React.Component {
         boxSizing: 'border-box'
       },
       Com_UnitViewSummary_Marksarticle: {
-        width: "44%",
+        width: "51%",
         height: '81.5%',
         position: 'absolute',
         right: '0%',
@@ -88,34 +75,10 @@ class UnitAuthorSummary extends React.Component {
     this.props._set_layerstatus(true, parseInt(moveCount), marksStatus);
   }
 
-  _handleClick_UnitAction_Author(event){
-    event.preventDefault();
-    event.stopPropagation();
-    this.props._set_Modalmode("author_editing");
-  }
-
   _handleClick_Account(event){
     event.preventDefault();
     event.stopPropagation();
     this.props._refer_toandclose('user', this.props.unitCurrent.authorBasic.authorId);
-  }
-
-  _handleEnter_actEdit(e){
-    //don't need to stop proppagation,
-    //because both the 'onMouseEnter' & 'onMouseLeave'
-    //would not 'bubble'
-    this.setState({
-      onActEdit: true
-    })
-  }
-
-  _handleLeave_actEdit(e){
-    //don't need to stop proppagation,
-    //because both the 'onMouseEnter' & 'onMouseLeave'
-    //would not 'bubble'
-    this.setState({
-      onActEdit: false
-    })
   }
 
   componentDidMount(){
@@ -173,22 +136,14 @@ class UnitAuthorSummary extends React.Component {
             _handleClick_listNoun={this.props._refer_toandclose}/>
         </div>
         <div
-          style={styleMiddle.boxStatics}>
-          <AuthorStatics/>
+          className={classnames()}>
+
+        </div>
+        <div
+          style={this.style.Com_UnitViewSummary_panel_}>
           <div
-            className={classnames('sumPanelOptions', 'boxSumOptEdit')}>
-            <span
-              className={classnames('fontSumOpt')}
-              style={Object.assign(
-                {},
-                {cursor: 'pointer'},
-                this.state.onActEdit? {color: '#FAFAFA'}:{color: 'rgba(250,250,250,0.5)'}
-              )}
-              onClick={this._handleClick_UnitAction_Author}
-              onMouseEnter={this._handleEnter_actEdit}
-              onMouseLeave={this._handleLeave_actEdit}>
-              {"edit"}
-            </span>
+            className={classnames()}>
+            <ActionBroad/>
           </div>
         </div>
       </div>
@@ -206,4 +161,4 @@ const mapStateToProps = (state)=>{
 export default withRouter(connect(
   mapStateToProps,
   null
-)(UnitAuthorSummary));
+)(UnitViewSummary));
